:art: (format): Change indentation style from tabs to spaces
We usually prefer tabs, but it turns out that using spaces will make our
lives easier when working with text based documents and variables such
as yaml and markdown.

Using tabs, swift-format struggles to correctly indent content which
results in a lot of time being spent to manually fix the issues which is
kind of a waste of our precious time.

So be it! Let's change tabs to spaces and never think about it ever
again... please!