🩹 (GEK): ObserveThen - Fix scaling issue, apply clipshape first
we display the image in a square of 460 by 460 pixels.
some images are rectangles, so we need to scale them to fit the frame.
fit is used instead of fill to make sure all the image is visible

we must apply the clipshape(RoundedRectangle) before the scaledToFit to
make sure it is visible.